暗黙で型変換される箇所から型変換の処理を削除

+++ DreamSat/Camera.cydsn/communicate_to_camera.c
@@ -93,7 +93,7 @@ void IsrCamRx() {
         case DATA:
             camera_buff[recv_length] = recv_data;
             ++recv_length;
-            if (recv_data_length <= (uint16)recv_length) {
+            if (recv_data_length <= recv_length) {
                 ++recv_phase;
             }
             break;
